+++ MMSApplication/src/main/java/com/javaschool/onlineshop/dao/ProductDAO.java
@@ -8,7 +8,7 @@ public interface ProductDAO {
 
     List<Product> findAll();
 
-    Product getProductById(int id);
+    Product getProductById(Long id);
 
     void addProduct(Product product);
